requirements.txt
discord.py
pandas
requests
python-dotenv
pyyaml
aiohttp
aiosqlite
pytest
pytest-asyncio